+++ util-core/src/main/java/com/twitter/util/Vars.java
@@ -34,8 +34,8 @@ public final class Vars {
   /**
    * @see com.twitter.util.Var$#value(Object)
    */
-  public static <T> Var<T> newConstVar(T constant) {
-    return Var$.MODULE$.value(constant);
+  public static <T> ConstVar<T> newConstVar(T constant) {
+    return new ConstVar<T>(constant);
   }
 
   /**