+++ util-core/src/main/scala-2.13+/com/twitter/conversions/SeqUtil.scala
@@ -1,0 +1,6 @@
+package com.twitter.conversions
+
+object SeqUtil {
+  def hasKnownSize[A](self: Seq[A]): Boolean =
+    self.knownSize != -1
+}